requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

python-dotenv==1.0.1
azure-functions==1.20.0
beautifulsoup4==4.12.3
instaloader==4.13.1
pandas==2.2.2
requests==2.32.3